rename offerings topic to supply
https://github.com/root-systems/cobuy/pull/65#discussion_r124711090
